refactor content-process for sfari - 1
