+++ package.json
@@ -97,7 +97,7 @@
     "vue-svg-loader": "^0.10.0",
     "vue-template-compiler": "^2.5.17",
     "webextension-polyfill": "^0.3.1",
-    "webpack": "^4.21.0",
+    "webpack": "^4.22.0",
     "webpack-bundle-analyzer": "^3.0.3",
     "webpack-dev-server": "^3.1.9",
     "webpack-merge": "^4.1.4",
